% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_graph.R
\name{create_function_graph}
\alias{create_function_graph}
\title{Create a graph that describes network of package functions}
\usage{
create_function_graph(pkg_fcn_info, pkg_name = NULL)
}
\arguments{
\item{pkg_fcn_info}{a tibble object that contains
package function information. This is created by the
\code{get_pkg_fcn_info()} function.}

\item{pkg_name}{an optional package name for filtering
the tibble provided to \code{pkg_fcn_info}, which is
useful if that tibble describes multiple packages.}
}
\description{
Create a DiagrammeR graph that contains nodes of
the type \code{function} and edges that have the
relationship \code{called_in}.
}
