% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_github.R
\name{from_github}
\alias{from_github}
\title{Helper function for pointing to an R package on GitHub}
\usage{
from_github(repo)
}
\arguments{
\item{repo}{The GitHub repository address of the following construction:
\code{[username]/[repo]}.}
}
\description{
This helper function allows for inspection of R packages that are hosted on
GitHub.
}
