% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_graph.R
\name{function_graph_single}
\alias{function_graph_single}
\title{Show all package functions that are called from a given function}
\usage{
function_graph_single(pkgattrs_tbl, target_fcn, pkg_name = NULL)
}
\arguments{
\item{pkgattrs_tbl}{A tibble object that contains package function
information. This is created by the \code{pkgattrs()} function.}

\item{target_fcn}{the name of the function that is to be examined for its
calls of package functions.}

\item{pkg_name}{An optional package name for filtering the tibble provided to
\code{pkgattrs_tbl}, which is useful if that tibble describes multiple
packages.}
}
\description{
Create a DiagrammeR graph that contains nodes of the type \code{function} and
edges that have the relationship \code{called_in}.
}
