% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fcn_lines_info.R
\name{get_fcn_lines_info}
\alias{get_fcn_lines_info}
\title{Get a table with a line-by-line info of package fcns}
\usage{
get_fcn_lines_info(..., .fcn_info_tbl = NULL, .make_clean = TRUE)
}
\arguments{
\item{...}{a series of objects pointing to
package locations. These can be strings with
paths to local package directories, or,
invocations of helper functions such as
\code{from_github()}.}

\item{.fcn_info_tbl}{an optional tibble of
function information obtained from the
\code{get_pk_fcn_info()} function.}

\item{.make_clean}{an option to clean the
working directory of any temporary package files
downloaded from GitHub.}
}
\description{
Create a tibble of information related to each
line of every function available in a package.
}
