% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkg_fcn_info.R
\name{get_pkg_fcn_info}
\alias{get_pkg_fcn_info}
\title{Get a table of info for a package's functions}
\usage{
get_pkg_fcn_info(...)
}
\arguments{
\item{...}{a series of objects pointing to
package locations. These can be strings with
paths to local package directories, or,
invocations of helper functions such as
\code{from_github()}.}
}
\description{
Create a tibble of information related to each
function available in a package.
}
