% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_fcns.R
\name{pkg_api}
\alias{pkg_api}
\title{Write the pkg API to disk}
\usage{
pkg_api(filename, ...)
}
\arguments{
\item{filename}{the output filename for the
text to be written to disk.}

\item{...}{any objects pointing to a package
location. This can be a string with a
path to a local package directory, or, an
invocation of a helper function such as
\code{from_github()}.}
}
\description{
Exported functions from a package (along with
their arguments and any associated default values)
are printed to a file.
}
