% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgattrs.R
\name{pkgattrs}
\alias{pkgattrs}
\title{Get a table of info of package fcns}
\usage{
pkgattrs(..., .make_clean = TRUE)
}
\arguments{
\item{...}{a series of objects pointing to package locations. These can be
strings with paths to local package directories, or, invocations of helper
functions such as \code{from_github()}.}

\item{.make_clean}{an option to clean the working directory of any temporary
package files downloaded from GitHub.}
}
\description{
Create a tibble of information related to each function available in a
package.
}
