% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_graph.R
\name{show_called_functions}
\alias{show_called_functions}
\title{Show all package functions that are called from a given function}
\usage{
show_called_functions(fcn_graph, caller_fcn)
}
\arguments{
\item{fcn_graph}{a function network graph generated by the
\code{create_function_graph()} function.}

\item{caller_fcn}{the name of the function that is
to examined for its calls of package functions.}
}
\description{
Create a DiagrammeR graph that contains nodes of
the type \code{function} and edges that have the
relationship \code{called_in}.
}
