% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pkg_api.R
\name{write_pkg_api}
\alias{write_pkg_api}
\title{Write the pkg API to disk}
\usage{
write_pkg_api(..., filename)
}
\arguments{
\item{...}{anf objects pointing to a package
locations. This can be a string with a
path to a local package directory, or, an
invocation of a helper functions such as
\code{from_github()}.}

\item{filename}{the output filename for the
text to be written to disk.}
}
\description{
Write the pkg API to disk
}
