% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pkg_fcn_info.R
\name{get_pkg_fcn_info}
\alias{get_pkg_fcn_info}
\title{Get a table of info for a package's functions}
\usage{
get_pkg_fcn_info(pkg_location)
}
\arguments{
\item{pkg_location}{the path to the root
directory of a package.}
}
\description{
Create a tibble of information related to each
function available in a package.
}
