% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgattrs.R
\name{pkgattrs}
\alias{pkgattrs}
\title{Get a table of info functions inside one or more packages}
\usage{
pkgattrs(..., .make_clean = TRUE, .get_cyclocomp = FALSE)
}
\arguments{
\item{...}{A series of objects pointing to package locations. These can be
strings with paths to local package directories, or, invocations of helper
functions such as \code{from_github()}.}

\item{.make_clean}{An option to clean the working directory of any temporary
package files downloaded from GitHub.}

\item{.get_cyclocomp}{An option to include a measure of each function's
cyclomatic complexity.}
}
\description{
Create a tibble of information related to each function available in one or
several packages.
}
