% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pkg_api.R
\name{write_pkg_api}
\alias{write_pkg_api}
\title{Write the pkg API to disk}
\usage{
write_pkg_api(..., filename)
}
\arguments{
\item{...}{Any objects pointing to a package location. This can be a string
with a path to a local package directory, or, an invocation of a helper
function such as \code{from_github()}.}

\item{filename}{The output filename for the text to be written to disk.}
}
\description{
Exported functions from a package (along with their arguments and any
associated default values) are printed to a file.
}
